+++ game/src/char.h
@@ -1107,7 +1107,7 @@ class CHARACTER : public CEntity, public CFSM, public CHorseRider
 
 		bool			RefineItem(LPITEM pkItem, LPITEM pkTarget);
 		bool			DestroyItem(TItemPos Cell);
-		bool			DropItem(TItemPos Cell,  BYTE bCount=0);
+		bool			DropItem(TItemPos Cell,  WORD bCount=0);
 		bool			GiveRecallItem(LPITEM item);
 		void			ProcessRecallItem(LPITEM item);
 
@@ -1131,7 +1131,7 @@ class CHARACTER : public CEntity, public CFSM, public CHorseRider
 		bool			GiveItemFromSpecialItemGroup(DWORD dwGroupNum, std::vector <DWORD> &dwItemVnums, 
 							std::vector <DWORD> &dwItemCounts, std::vector <LPITEM> &item_gets, int &count);
 
-		bool			MoveItem(TItemPos pos, TItemPos change_pos, BYTE num);
+		bool			MoveItem(TItemPos pos, TItemPos change_pos, WORD num);
 		bool			PickupItem(DWORD vid);
 		bool			EquipItem(LPITEM item, int iCandidateCell = -1);
 		bool			UnequipItem(LPITEM item);
@@ -1143,7 +1143,7 @@ class CHARACTER : public CEntity, public CFSM, public CHorseRider
 		bool			CanUnequipNow(const LPITEM item, const TItemPos& srcCell = NPOS, const TItemPos& destCell = NPOS);
 
 		bool			SwapItem(BYTE bCell, BYTE bDestCell);
-		LPITEM			AutoGiveItem(DWORD dwItemVnum, BYTE bCount=1, int iRarePct = -1, bool bMsg = true);
+		LPITEM			AutoGiveItem(DWORD dwItemVnum, WORD bCount=1, int iRarePct = -1, bool bMsg = true);
 		void			AutoGiveItem(LPITEM item, bool longOwnerShip = false);
 		
 		int				GetEmptyInventory(BYTE size) const;
